Add scene part loading from OBJ files
